[Feat]🥚관리자 대면상담-문의 정보 API 구현

+++ src/main/java/com/hanaro/schedule_hanaro/customer/repository/CsVisitRepository.java
@@ -14,7 +14,7 @@ import com.hanaro.schedule_hanaro.global.domain.CsVisit;
 import jakarta.persistence.LockModeType;
 
 @Repository
-public interface CsVisitRepository extends JpaRepository<CsVisit, Integer> {
+public interface CsVisitRepository extends JpaRepository<CsVisit, Long> {
 
 	Optional<CsVisit> findByBranchIdAndDate(Long branchId, LocalDate date);
